metrics.rs: Refactor dashboard rendering to its own module
