+++ Cargo.toml
@@ -44,6 +44,8 @@ serde-this-or-that = "0.4.0"
 # Rationale, 2023-03-21: https://stackoverflow.com/questions/74462753
 typed-html = { git = "https://github.com/bodil/typed-html", rev = "4c13ecca" }
 humantime = "2.1.0"
+prometheus = "0.13.3"
+lazy_static = "1.4.0"
 
 [dev-dependencies]
 tokio-util = { version = "0.7.0", features = ["full"] }
