+++ src/agent.rs
@@ -46,6 +46,7 @@ Note that there is an Oracle and Exporter for each network, but only one Local S
 
 ################################################################################################################################## */
 
+pub mod dashboard;
 pub mod metrics;
 pub mod pythd;
 pub mod remote_keypair_loader;
